+++ test/lz4_test.dart
@@ -3,6 +3,7 @@
 // a BSD-style license that can be found in the LICENSE file.
 
 import 'dart:convert';
+import 'dart:io';
 
 import 'package:collection/collection.dart';
 import 'package:es_compression/lz4.dart';
@@ -73,6 +74,17 @@ void main() {
     final encoded = codec.encode(dataBytes);
     expect(const ListEquality<int>().equals(encoded, expected), true);
     final decoded = lz4.decode(encoded);
+    expect(const ListEquality<int>().equals(dataBytes, decoded), true);
+  });
+
+  test('Test Large Lz4 Encode/Decode', () async {
+    const filePath = 'test/data/256KB.json';
+    final file = File(filePath);
+    final data = await file.readAsString();
+    final dataBytes = utf8.encode(data);
+    final codec = Lz4Codec();
+    final encoded = codec.encode(dataBytes);
+    final decoded = codec.decode(encoded);
     expect(const ListEquality<int>().equals(dataBytes, decoded), true);
   });
 }